Extract Tip components into separate files
