+++ frontend/src/components/dashboard/tips/Tips.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 @import "~@mozilla-protocol/core/protocol/css/includes/lib";
 
 .wrapper {
@@ -170,39 +170,5 @@
         font-weight: 500;
       }
     }
-  }
-}
-
-.generic-tip {
-  @include text-body-sm;
-  display: flex;
-  flex-direction: column;
-  gap: $spacing-sm;
-
-  h3 {
-    @include text-title-3xs;
-    font-family: $font-stack-firefox;
-    color: $color-purple-90;
-  }
-}
-
-.custom-alias-tip {
-  display: flex;
-  flex-direction: column;
-  gap: $spacing-xs;
-
-  @include text-body-sm;
-
-  samp {
-    @include text-body-sm;
-    font-family: $font-stack-base;
-    color: $color-blue-50;
-    font-weight: 700;
-  }
-
-  h3 {
-    @include text-title-3xs;
-    font-family: $font-stack-firefox;
-    color: $color-purple-90;
   }
 }
